Refactor: organizes tests by feature and add to readme
